refactor: isolate connection error test support
