+++ src/app/model/connection/error_test_support.rs
@@ -1,0 +1,10 @@
+use super::ConnectionErrorInfo;
+
+pub fn from_parts(
+    summary: &'static str,
+    hint: &'static str,
+    retryable: bool,
+    raw_stderr: impl Into<String>,
+) -> ConnectionErrorInfo {
+    ConnectionErrorInfo::from_presentation(summary, hint, retryable, raw_stderr)
+}
